script to bulk rename PICO-8 roms
The two main extension types for PICO-8 roms are `.p8` and `.p8.png`, but the latter compound extension is supported by SimpleMenu. 

For the `alias_PICO-8.txt` to operate for both extensions `.p8.png` needs to be renamed as simply `.png`.

Also the preview images within SimpleMenu are loaded from different locations depending on the file extension.  For `.p8` the standard `.previews` folder is used for the images.  For `.png` the rom's in-built "cart" image is used.

This script provides an easy way for users to synchronise their PICO-8 rom collection so alias file is used and also allows them to toggle between the image preview types.